requirements.txt
numpy==1.21.6
Pillow==9.2.0
torch==1.12.1
